requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
langchain>=0.3.1
langchain-openai>=0.3.1
mcp-use>=0.1.0
pydantic>=2.10.6
fastapi
fastapi-mcp
mmh3
pymilvus>=2.4.0
openai>=1.30.0
python-docx>=0.8.11
openpyxl>=3.1.2
pypdf>=4.2.0
reportlab>=4.2.5
supabase>=2.6.0
